# documentation/api-spec.yaml
openapi: 3.0.0
info:
  title: "Deadline Tracker API"
  version: "1.0.0"

tags:
  - name: "schedule"
    description: "Interact with your schedule"

paths: 
  /schedule:
    get:
      tags: 
        - "schedule"
      summary: "Get all tasks"
      description: "Returns a list of all tasks"
      responses:
        "200":
          description: "List of all tasks"

  /schedule/{taskId}:
    get:
      tags: 
        - "schedule"
      summary: "Get a task by ID"
      description: "Returns a single task by ID"
      parameters:
        - name: "taskId"
          description: "The UID of the task"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "Details about a task"
        "404":
          description: "Task not found"

  /schedule/add:
    post:
      tags: 
        - "schedule"
      summary: "Add a task"
      description: "Creates a new task"
      parameters:
      - name: "taskName"
        description: "The name of the task "
        in: "query"
        required: true
        schema:
          type: "string"
      - name: "dueDate"
        description: "The date that the task is due"
        in: "query"
        required: true
        schema:
          type: "string"
      - name: "category"
        description: "The category the task belongs to (quiz, project, etc.)"
        in: "query"
        required: false
        schema:
          type: "string"
      - name: "priority"
        description: "The priority level of the task (Normal, High, Critical)"
        in: "query"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "Task added successfully"
        "400":
          description: "Invalid values submitted"
      
  /schedule/edit/{taskId}:
    put:
      tags: 
        - "schedule"
      summary: "Edit a task"
      description: "Updates info for an existing task by id"
      parameters:
        - name: "taskId"
          description: "The UID of the task"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "taskName"
          description: "The name of the task "
          in: "query"
          required: false
          schema:
            type: "string"
        - name: "dueDate"
          description: "The date that the task is due"
          in: "query"
          required: false
          schema:
            type: "string"
        - name: "category"
          description: "The category the task belongs to (quiz, project, etc.)"
          in: "query"
          required: false
          schema:
            type: "string"
        - name: "priority"
          description: "The priority level of the task (Normal, High, Critical)"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "Task updated successfully"
        "400":
          description: "Invalid values submitted"
        "404":
          description: "Task not found"

  /schedule/delete/{taskId}:
    delete:
      tags: 
        - "schedule"
      summary: "Delete a task"
      description: "Deletes a task by id"
      parameters:
        - name: "taskId"
          description: "The UID of the task"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "Task deleted successfully"
        "404":
          description: "Task not found"